backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
django-filter==24.2.0
PyJWT==2.9.0